refactor(core): move workspace search owner
Move local flashgrep workspace search service ownership into services-integrations behind the workspace-search feature while keeping bitfun-core as the compatibility facade for config/bootstrap hooks and legacy error mapping.

Move tool-call truncation recovery presentation into tool contracts and update boundary checks plus plan docs for the remaining remote search/tool orchestration work.
