+++ scripts/core-boundaries/rules/feature-rules.mjs
@@ -51,13 +51,13 @@ export const optionalDependencyFeatureOwnerRules = [
     dependencies: [
       { depName: 'aes-gcm', ownerFeatures: ['mcp'] },
       { depName: 'anyhow', ownerFeatures: ['mcp'] },
-      { depName: 'async-trait', ownerFeatures: ['mcp', 'remote-connect'] },
+      { depName: 'async-trait', ownerFeatures: ['mcp', 'remote-connect', 'workspace-search'] },
       { depName: 'base64', ownerFeatures: ['mcp', 'remote-connect'] },
       { depName: 'bitfun-product-domains', ownerFeatures: ['function-agents'] },
       { depName: 'bitfun-runtime-ports', ownerFeatures: ['remote-connect'] },
-      { depName: 'bitfun-services-core', ownerFeatures: ['git', 'mcp'] },
+      { depName: 'bitfun-services-core', ownerFeatures: ['git', 'mcp', 'workspace-search'] },
       { depName: 'chrono', ownerFeatures: ['git'] },
-      { depName: 'dunce', ownerFeatures: ['remote-ssh'] },
+      { depName: 'dunce', ownerFeatures: ['remote-ssh', 'workspace-search'] },
       { depName: 'futures', ownerFeatures: ['mcp'] },
       { depName: 'git2', ownerFeatures: ['git'] },
       { depName: 'notify', ownerFeatures: ['file-watch'] },
@@ -66,9 +66,10 @@ export const optionalDependencyFeatureOwnerRules = [
       { depName: 'rmcp', ownerFeatures: ['mcp'] },
       { depName: 'sha2', ownerFeatures: ['remote-ssh'] },
       { depName: 'sse-stream', ownerFeatures: ['mcp'] },
-      { depName: 'thiserror', ownerFeatures: ['git'] },
+      { depName: 'thiserror', ownerFeatures: ['git', 'workspace-search'] },
       { depName: 'tokio-util', ownerFeatures: ['remote-ssh'] },
       { depName: 'uuid', ownerFeatures: ['remote-connect'] },
+      { depName: 'which', ownerFeatures: ['workspace-search'] },
     ],
   },
   {
@@ -149,6 +150,7 @@ export const ownerCrateFeatureAssemblyRules = [
       'mcp',
       'remote-connect',
       'remote-ssh',
+      'workspace-search',
     ],
   },
   {
