+++ src/crates/services/services-integrations/src/workspace_search/flashgrep/types.rs
@@ -1,11 +1,11 @@
-pub(crate) use super::protocol::{
+pub use super::protocol::{
     ConsistencyMode, DirtyFileStats, FileCount, OpenRepoParams, PathScope, QuerySpec,
     RefreshPolicyConfig, RepoConfig, RepoPhase, RepoStatus, SearchBackend, SearchModeConfig,
     SearchResults, TaskKind, TaskPhase, TaskState, TaskStatus, WorkspaceOverlayStatus,
 };
 
 #[derive(Debug, Clone)]
-pub(crate) struct SearchRequest {
+pub struct SearchRequest {
     pub query: QuerySpec,
     pub scope: PathScope,
     pub consistency: ConsistencyMode,
@@ -13,25 +13,25 @@ pub(crate) struct SearchRequest {
 }
 
 #[derive(Debug, Clone, Default)]
-pub(crate) struct GlobRequest {
+pub struct GlobRequest {
     pub scope: PathScope,
 }
 
 #[derive(Debug, Clone)]
-pub(crate) struct SearchOutcome {
+pub struct SearchOutcome {
     pub backend: SearchBackend,
     pub status: RepoStatus,
     pub results: SearchResults,
 }
 
 #[derive(Debug, Clone)]
-pub(crate) struct GlobOutcome {
+pub struct GlobOutcome {
     pub status: RepoStatus,
     pub paths: Vec<String>,
 }
 
 impl SearchRequest {
-    pub(crate) fn new(query: QuerySpec) -> Self {
+    pub fn new(query: QuerySpec) -> Self {
         Self {
             query,
             scope: PathScope::default(),
@@ -40,28 +40,28 @@ impl SearchRequest {
         }
     }
 
-    pub(crate) fn with_scope(mut self, scope: PathScope) -> Self {
+    pub fn with_scope(mut self, scope: PathScope) -> Self {
         self.scope = scope;
         self
     }
 
-    pub(crate) fn with_consistency(mut self, consistency: ConsistencyMode) -> Self {
+    pub fn with_consistency(mut self, consistency: ConsistencyMode) -> Self {
         self.consistency = consistency;
         self
     }
 
-    pub(crate) fn with_scan_fallback(mut self, allow_scan_fallback: bool) -> Self {
+    pub fn with_scan_fallback(mut self, allow_scan_fallback: bool) -> Self {
         self.allow_scan_fallback = allow_scan_fallback;
         self
     }
 }
 
 impl GlobRequest {
-    pub(crate) fn new() -> Self {
+    pub fn new() -> Self {
         Self::default()
     }
 
-    pub(crate) fn with_scope(mut self, scope: PathScope) -> Self {
+    pub fn with_scope(mut self, scope: PathScope) -> Self {
         self.scope = scope;
         self
     }
